+++ src/DomHandler/AbstractElementHandler.php
@@ -47,6 +47,19 @@ abstract class AbstractElementHandler extends AbstractNodeHandler
     }
     /**
      * @param string $name
+     * @return mixed
+     */
+    public function getAttributeValue($name, $withNamespace = false, $withinItsType = true, $asType = AbstractAttributeHandler::DEFAULT_VALUE_TYPE)
+    {
+        $value = null;
+        $attribute = $this->getAttribute($name);
+        if ($attribute instanceof AbstractAttributeHandler) {
+            $value = $attribute->getValue($withNamespace, $withinItsType, $asType);
+        }
+        return $value;
+    }
+    /**
+     * @param string $name
      * @return NodeHandler[]|ElementHandler[]
      */
     public function getChildrenByName($name)
@@ -88,5 +101,64 @@ abstract class AbstractElementHandler extends AbstractNodeHandler
     {
         $children = $this->getChildrenByNameAndAttributes($name, $attributes);
         return empty($children) ? null : array_shift($children);
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return mixed
+     */
+    public function getMaxOccurs()
+    {
+        $maxOccurs = $this->getAttributeValue(AbstractAttributeHandler::ATTRIBUTE_MAX_OCCURS);
+        if ($maxOccurs === AbstractAttributeHandler::VALUE_UNBOUNDED) {
+            return $maxOccurs;
+        }
+        if (!is_numeric($maxOccurs)) {
+            return AbstractAttributeHandler::DEFAULT_OCCURENCE_VALUE;
+        }
+        return (int)$maxOccurs;
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return int
+     */
+    public function getMinOccurs()
+    {
+        $minOccurs = $this->getAttributeValue(AbstractAttributeHandler::ATTRIBUTE_MIN_OCCURS);
+        if (!is_numeric($minOccurs)) {
+            return AbstractAttributeHandler::DEFAULT_OCCURENCE_VALUE;
+        }
+        return (int)$minOccurs;
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return bool
+     */
+    public function canOccurSeveralTimes()
+    {
+        return ($this->getMinOccurs() > 1) || ($this->getMaxOccurs() > 1) || ($this->getMaxOccurs() === AbstractAttributeHandler::VALUE_UNBOUNDED);
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return bool
+     */
+    public function canOccurOnlyOnce()
+    {
+        return $this->getMaxOccurs() === 1;
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return bool
+     */
+    public function isOptional()
+    {
+        return $this->getMinOccurs() === 0;
+    }
+    /**
+     * Info at {@link https://www.w3.org/TR/xmlschema-0/#OccurrenceConstraints}
+     * @return bool
+     */
+    public function isRequired()
+    {
+        return $this->getMinOccurs() >= 1;
     }
 }