+++ tests/DomHandler/DomDocumentHandlerTest.php
@@ -10,6 +10,7 @@ class DomDocumentHandlerTest extends TestCase
     protected static $ebayInstance;
     protected static $bingInstance;
     protected static $emptyInstance;
+    protected static $yandexDirectApiAdGroupsInstance;
     /**
      * @return DomDocumentHandler
      */
@@ -17,7 +18,7 @@ class DomDocumentHandlerTest extends TestCase
     {
         if (!isset(self::$ebayInstance)) {
             $doc = new \DOMDocument('1.0', 'utf-8');
-            $doc->load(__DIR__ . '/../resources/ebaySvc.wsdl');
+            $doc->load(self::wsdlEbayPath());
             self::$ebayInstance = new DomDocumentHandler($doc);
         }
         return self::$ebayInstance;
@@ -29,7 +30,7 @@ class DomDocumentHandlerTest extends TestCase
     {
         if (!isset(self::$bingInstance)) {
             $doc = new \DOMDocument('1.0', 'utf-8');
-            $doc->load(__DIR__ . '/../resources/bingsearch.wsdl');
+            $doc->load(self::wsdlBingPath());
             self::$bingInstance = new DomDocumentHandler($doc);
         }
         return self::$bingInstance;
@@ -41,10 +42,22 @@ class DomDocumentHandlerTest extends TestCase
     {
         if (!isset(self::$emptyInstance)) {
             $doc = new \DOMDocument('1.0', 'utf-8');
-            @$doc->load(__DIR__ . '/../resources/empty.wsdl');
+            @$doc->load(self::wsdlEmptyPath());
             self::$emptyInstance = new DomDocumentHandler($doc);
         }
         return self::$emptyInstance;
+    }
+    /**
+     * @return DomDocumentHandler
+     */
+    public static function yandeDirectApiAdGroupsInstance()
+    {
+        if (!isset(self::$yandexDirectApiAdGroupsInstance)) {
+            $doc = new \DOMDocument('1.0', 'utf-8');
+            $doc->load(self::wsdlYandexDirectApiAdGroupsPath());
+            self::$yandexDirectApiAdGroupsInstance = new DomDocumentHandler($doc);
+        }
+        return self::$yandexDirectApiAdGroupsInstance;
     }
     /**
      *