+++ tests/File/StructTest.php
@@ -153,6 +153,38 @@ class StructTest extends AbstractFile
     /**
      *
      */
+    public function testWriteYandexDirectApiStructAddRequest()
+    {
+        $generator = self::yandexDirectApiAdGroupsGeneratorInstance();
+        if (($model = $generator->getStruct('AddRequest')) instanceof StructModel) {
+            $struct = new StructFile($generator, $model->getName());
+            $struct
+                ->setModel($model)
+                ->write();
+            $this->assertSameFileContent('ValidAddRequest', $struct);
+        } else {
+            $this->assertFalse(true, 'Unable to find AddRequest struct for file generation');
+        }
+    }
+    /**
+     *
+     */
+    public function testWriteYandexDirectApiStructAdGroupsSelectionCriteria()
+    {
+        $generator = self::yandexDirectApiAdGroupsGeneratorInstance();
+        if (($model = $generator->getStruct('AdGroupsSelectionCriteria')) instanceof StructModel) {
+            $struct = new StructFile($generator, $model->getName());
+            $struct
+                ->setModel($model)
+                ->write();
+            $this->assertSameFileContent('ValidAdGroupsSelectionCriteria', $struct);
+        } else {
+            $this->assertFalse(true, 'Unable to find AdGroupsSelectionCriteria struct for file generation');
+        }
+    }
+    /**
+     *
+     */
     public function testDestination()
     {
         $generator = self::bingGeneratorInstance();